requirements.txt
Django==5.1.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
djangorestframework-simplejwt==5.3.1
google-generativeai==0.3.2
Pillow==10.1.0
drf-yasg==1.21.7
whitenoise==6.6.0
gunicorn==21.2.0
python-multipart==0.0.6
